requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
openai==1.52.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pypdfium2==4.30.0
python-dotenv==1.0.1
slack_sdk==3.33.2
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
